Move final layout into prop function
We move the final layout of the tags & input into a function provided by
the props with a default implementation.

The default reproduces the current layout but the mechanism allows
client code to override the final layout so that, for example, the input
can appear above the tags instead of after them.

We use function.call to pass the 'this' down to the layout function so
that it can access this.props and generally be an extension of the
render function.